+++ pages/multi-complete/index.html
@@ -25,7 +25,7 @@
                 <button>Меню</button>
         </div>
     </div>
-    <!-- <br> -->
+
     <div class="grid">
         <div class="item item_zero">
             <div class="zero">0
