+++ pages/multi-complete/style.css
@@ -123,3 +123,4 @@ body {
 .vertical-text {
     writing-mode: vertical-rl;
 }
+
